FEAT: New File generation

+++ modeling/src/schedule.rs
@@ -2,11 +2,11 @@ use shared::project::Project;
 
 pub struct Schedule<'a> {
     pub(crate) projects: Vec<&'a Project<'a>>,
-    pub(crate) resources: Vec<u32>,
+    pub(crate) resources: Vec<usize>,
 }
 
 impl<'a> Schedule<'a> {
-    pub fn new(projects: Vec<&'a Project<'a>>, resources: Vec<u32>) -> Self {
+    pub fn new(projects: Vec<&'a Project<'a>>, resources: Vec<usize>) -> Self {
         Self {
             projects,
             resources,
